commit split and encoder
